requirements.txt
Flask==3.0.0
gunicorn==21.0.1
langchain==0.0.339
python-dotenv==1.0.0
openai==1.2.3
todoist_api_python==2.1.3